fix(cross-compile): windows GOOS build failures (syscall.Kill, syscall.Statfs)
GoReleaser's windows/amd64 target has never actually been built until the
first real tag push (v1.3.0) exercised it — PRD 38 shipped without
ever running an end-to-end cross-compile check for that target.

- internal/adapters/mongo_tls: processAlive split into
  process_unix.go (linux || darwin, syscall.Kill-based) and
  process_other.go (conservative "always alive" stub — SweepOrphanMaterial
  just skips cleanup on non-POSIX rather than failing to build).
- internal/adapters/restore/runtime: staging capacity check split into
  diskspace_unix.go (syscall.Statfs-based) and diskspace_other.go (skips
  the advisory pre-flight check on non-POSIX rather than failing to build
  or blocking restores).

go build/vet/test clean natively and cross-compiled for all 5 GoReleaser
targets (linux/darwin amd64+arm64, windows amd64).
